feat: Add data mocks | Update experiences
* Add data mocks
* Update experiences
